[NEW] initiate db once new guildMember
