+++ db/db.json.example
@@ -1,4 +1,5 @@
 {
+  "alavirien": [],
   "apologiesCounting": [],
   "birthdaysUsers": [],
   "concreteRecipients": [],
